+++ src/ptr_weak_weak_hash_map.rs
@@ -210,6 +210,7 @@ where
     }
 }
 
+#[cfg(any(test, feature = "std", feature = "ahash"))]
 impl<K, V, const N: usize> From<[(K::Strong, V::Strong); N]>
     for PtrWeakWeakHashMap<K, V, RandomState>
 where
@@ -277,6 +278,8 @@ impl<'a, K: WeakElement, V: WeakElement, S> IntoIterator for &'a PtrWeakWeakHash
 #[cfg(test)]
 mod test {
     #![allow(clippy::print_stderr)]
+    // TODO 050: remove.
+    #![cfg_attr(feature = "ahash", allow(deprecated))]
 
     use super::{Entry, PtrWeakWeakHashMap};
     use crate::{