+++ CHANGELOG.md
@@ -48,6 +48,10 @@ The format is based on [Keep a Changelog] and this project adheres to
   - `From<[T;N]>`
   - `Sub`
 
+- It is now possible to build as `no_std` without ahash support.
+  If you do, there will be no default `BuildHasher`,
+  and the `new()` and `with_capacity()` methods will be absent.
+
 ### Changed (visible)
 
 - All hash-tables now use a new backend based on [`hashbrown`]
@@ -60,7 +64,9 @@ The format is based on [Keep a Changelog] and this project adheres to
   actual load factor.
 - Methods that do not require the element type to be WeakKey or
   WeakElement no longer constrain their parameters in this way.
-
+- The `ahash` feature is now deprecated for [security reasons](github#23).
+  Instead, use the constructors that allow you to specify
+  a `BuildHasher` manually.
 
 ### Changed (internal)
 
@@ -184,3 +190,4 @@ Initial release.
 [`weak-table`]: <https://crates.io/crates/weak-table>
 [`hashbrown`]: <https://docs.rs/hashbrown/latest/hashbrown/struct.HashTable.html>
 [github#22]: <https://github.com/tov/weak-table-rs/issues/22>
+[github#23]: <https://github.com/tov/weak-table-rs/issues/23>